[ADD] event_product: extract product from sales related module *
*: event_booth_sale, event_sale, website_event_sale

Before product integration was directly managed in event_sale module.

This commit extract product management in a new module event_product.

taskId: 2995558
