+++ addons/event_product/models/__init__.py
@@ -1,0 +1,3 @@
+from . import event_event
+from . import event_ticket
+from . import product
