+++ addons/event_product/models/product.py
@@ -1,0 +1,17 @@
+# -*- coding: utf-8 -*-
+
+from odoo import fields, models
+
+
+class ProductTemplate(models.Model):
+    _inherit = 'product.template'
+
+    service_tracking = fields.Selection(selection_add=[
+        ('event', 'Event Registration'),
+    ], ondelete={'event': 'set default'})
+
+
+class Product(models.Model):
+    _inherit = 'product.product'
+
+    event_ticket_ids = fields.One2many('event.event.ticket', 'product_id', string='Event Tickets')
