+++ addons/event_sale/models/__init__.py
@@ -1,7 +1,5 @@
 from . import event_event
 from . import event_registration
 from . import event_ticket
-from . import product_product
-from . import product_template
 from . import sale_order
 from . import sale_order_line
